add HasEnemyHitbox trait
