+++ rust/src/traits/components/character_components/mod.rs
@@ -1,5 +1,6 @@
 pub mod character_resources;
 pub mod damageable;
 pub mod damaging;
+pub mod has_hitbox;
 pub mod moveable;
 pub mod player;
